# Fix images on CODE